# docs/api.yaml
openapi: 3.0.2
info:
    title: Bank
    version: 1.0.0
    description: Markko Poljakov ja Emil Värnomasing BANK
    contact:
        name: Markko Poljakov ja Emil Värnomasing
        email: markko.poljakov@khk.ee
paths:
...
    /api/session:
        post:
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/postSessionRequest'
                required: true
            tags:
                - Sessions
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/postSessionResponse'
                            examples:
                                tokenHasBeenCreated:
                                    value:
                                        token: 5f9db108b8c28c40a42d94c4
                    description: ''
                '400':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/postSessionResponse'
                            examples:
                                userAndPwRequired:
                                    value:
                                        error: Username and password required
                                invalidUsernameOrPassword:
                                    value:
                                        error: Invalid username or password
                    description: Invalid username or password
                '500':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/postSessionResponse'
                            examples:
                                oopsSomethingWentWrong:
                                    value:
                                        error: Oops! Something went wrong
                    description: Oops! Something went wrong
            summary: Create session token
            description: Create session token
        delete:
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/json'
                required: true
            tags:
                - Sessions
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/postSessionResponse'
                            examples:
                                deleteToken:
                                    value:
                                        message: Token successfully deleted
                    description: Deleted token
                '401':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/errorResponse'
                            examples:
                                "missingAuthorizationHeader\t":
                                    value: "\t{\"error\":\"Missing Authorization header\"}"
                                invalidHeaderFormat:
                                    value: "\t{\"error\":\"Invalid Authorization header format\"}"
                                invalidToken:
                                    value: "\t{\"error\":\"Invalid token\"}"
                                invalidMongooseToken:
                                    value:
                                        error: Invalid token
                    description: Missing token
                '404':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/errorResponse'
                            examples:
                                sessionNotFound:
                                    value:
                                        error: Session not found
                    description: Token not found
                '500':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/errorResponse'
                            examples:
                                oopsSomethingWentWrong:
                                    value: "\t{\"error\":\"Oops! Something went wrong\"}"
                    description: Server error
            security:
                -
                    Bearer: []
            summary: Delete session token
            description: Delete session token
components:
    schemas:
...
        errorResponse:
            title: Error response
            description: Generic error response
            required:
                - missingAuthorizationHeader
                - error
                - invalidHeaderFormat
                - invalidToken
                - invalidMongooseToken
                - insufficentFunds
                - invalidAmount
                - invalidAccountTo
                - cannotTransferToYourself
                - noLoggedTransactions
            type: object
            properties:
                error:
                    description: Error response
                    type: string
                missingAuthorizationHeader:
                    description: '{"error":"Missing Authorization header"}'
                    type: string
                invalidHeaderFormat:
                    description: "\t{\"error\":\"Invalid Authorization header format\"}"
                    type: string
                invalidToken:
                    description: '{"error":"Invalid token"}'
                    type: string
                invalidMongooseToken:
                    description: '{"error":"Invalid token"}'
                    type: string
                insufficentFunds:
                    description: '{ "error": ''Insufficient funds'' }'
                    type: string
                invalidAmount:
                    description: '{ "error": ''Invalid amount'' }'
                    type: string
                invalidAccountTo:
                    description: '{ "error": ''Invalid accountTo'' }'
                    type: string
                cannotTransferToYourself:
                    description: '{"error": "You cannot transfer to yourself"}'
                    type: string
                noLoggedTransactions:
                    description: '{ "error": "You have no logged transactions" }'
                    type: string
            example:
                error: error message
        postSessionResponse:
            description: ''
            type: object
            example:
                token: 5f6db108b8c28c40a43d94c4
        postSessionRequest:
            description: ''
            required:
                - username
                - password
            type: object
            properties:
                username:
                    description: ''
                    type: string
                password:
                    description: ''
                    type: string
            example:
                username: random text
                password: random text
...
        json:
            title: Root Type for json
            description: '{}'
            type: object
            example: {}
    securitySchemes:
        Bearer:
            scheme: bearer
            type: http